requirements.txt
gast==0.2.2
h5py==2.10.0
matplotlib>=3.3.2
numpy>=1.19.2
Pillow>=7.2.0
pyzmq>=19.0.1
requests>=2.18.4
ruamel.yaml>=0.16.12
stable-baselines
torch==1.7.1
torchvision
websocket_client==0.57.0
tqdm
